refactor: implementing new check email page
